gui/settings: move all themes part inside new themes file. - fix clock/date pos on start screen in second date layout.
